refactor(cs): remove gemini-history prefix from tracker files and update imports
The gemini-history- prefix was redundant as these files are already within the gemini-tracker directory. This change cleans up the file naming convention and simplifies imports across the tracker modules.
