+++ src/content-scripts/gemini-tracker/config.js
