+++ src/content-scripts/gemini-tracker/event-handlers.js
@@ -1,10 +1,10 @@
-import { STATE } from "./gemini-history-state.js";
-import { Utils } from "./gemini-history-utils.js";
-import { StatusIndicator } from "./gemini-history-status-indicator.js";
-import { DomObserver } from "./observer/gemini-history-dom-observer.js";
-import { GemDetector } from "./gemini-history-gem-detector.js";
-import { ModelDetector } from "./gemini-history-model-detector.js";
-import { InputExtractor } from "./gemini-history-input-extractor.js";
+import { STATE } from "./state.js";
+import { Utils } from "./utils.js";
+import { StatusIndicator } from "./status-indicator.js";
+import { DomObserver } from "./observer/dom-observer.js";
+import { GemDetector } from "./gem-detector.js";
+import { ModelDetector } from "./model-detector.js";
+import { InputExtractor } from "./input-extractor.js";
 
 export const EventHandlers = {
   /**
