+++ src/content-scripts/gemini-tracker/state.js
